add stream event handler
